Add a parameter "charging" and "percentagechanged" which lets us veto some
* src/gpm-marshal.list, src/gpm-manager.c, src/gpm-power.c:
Add a parameter "charging" and "percentagechanged" which lets us veto some actions when doing the initial startup. This stops the message "Your battery is fully charged"	when starting g-p-m with a full battery. Also unconditionally send a signal battery-power-changed when *any* battery property changes, so that the icon changes when !charging as this was not being done.

* MAINTAINERS: I have a name, I am not an email address.

* CodingStyle: Update with suggestions from (#326515)