Change dbus Inhibit and UnInhibit methods to use a UINT32, as this is
2006-06-14  Richard Hughes  <richard@hughsie.com>

* src/gpm-manager.xml:
* docs/dbus-interface.xml:
* docs/gnome-power-manager.html:
Change dbus Inhibit and UnInhibit methods to use a UINT32, as this is
already specified in the spec. Using a INT32 is wrong.

* data/Makefile.am:
* data/gpm-inhibit-test.glade:
* src/Makefile.am:
* src/gpm-inhibit-test.c:
Add a test program that allows us to test the multiple inhibit stuff,
and also to show people how easy it is adding Inhibit and UnInhibit
calls into an existing program.