Use the same logic as the tooltip for the notification. This has caused a
2006-06-23  Richard Hughes  <richard@hughsie.com>

* src/gpm-manager.c: (battery_status_changed_primary):
Use the same logic as the tooltip for the notification.
This has caused a lot of problems, see #345698 for all the gory details.
This means if the policy is set to 'show icon when charging' then the
icon removes itself and the 'fully charged' notification is shown in
the bottom right of the screen, which I think is sane.

* help/C/gnome-power-manager.xml:
Add a note mentioning the battery has to go down to 95% before a repeat
notification will be shown.

* src/gpm-hal-monitor.c: (watch_device_property_modified),
(watch_device_properties_modified):
* src/gpm-hal-monitor.h:
Add a boolean 'finally' for the battery_property_modified so we know
when a split PropertyModified HAL event has finished, and we can update
the caches and recompute only when we have all the information.
This seems like a pain in the arse, but actually reduces our CPU usage
by nearly *5* times on a typical /org/freedesktop/Hal/devices/acpi_BAT1
type event update so was worth doing.

* src/gpm-power.c: (hal_battery_property_modified_cb):
Only refresh the caches on the last PropertyModified split event.