Don't poke the screen when we shut the lid. This should stop the display
2006-10-17  Richard Hughes <richard@hughsie.com>

* src/gpm-manager.c: (power_button_pressed_cb):
Don't poke the screen when we shut the lid. This should stop the
display going off, on, off when the lid is closed. Should fix #362694.

* src/gpm-power.c: (battery_kind_cache_update):
When we fall back to using the ac-adapter, we check the percentage.
The percentage was set below the check, and thus always failed.
Pointed out in #361583, and will probably affect multibattery laptops.