Add a couple of modules to the prefs program for deps.
2006-10-28  Richard Hughes <richard@hughsie.com>

* src/Makefile.am:
Add a couple of modules to the prefs program for deps.

* po/POTFILES.in:
* src/gpm-brightness-kbd.c:
* src/gpm-brightness-kbd.h:
* src/gpm-brightness-lcd.c:
* src/gpm-brightness-lcd.h:
* src/gpm-cpufreq.c:
* src/gpm-cpufreq.h:
* src/gpm-inhibit.c:
* src/gpm-light-sensor.c:
* src/gpm-light-sensor.h:
* src/gpm-prefs-core.c:
Remove the HAL prefix from these classes.

* src/gpm-manager.c:
Move some of the state functionality from the manager into the
objects. The manager is a mess and stuff needs to be better contained
in thier own classes.
The idea is that a class, say for instance the cpufreq class, is aware
of what to do on ac-state-change and does it itself. Therefore all the
functionality is in this module rather than half in the module and half
in the manager.
All the multiply used classes are singletons and so we don't use any
more memory.