Use GpmControl to notify GpmProfile when a suspend is about to happen, or
2007-07-05  Richard Hughes  <richard@hughsie.com>

* src/gpm-profile.c: (control_sleep_action_cb), (gpm_profile_init),
(gpm_profile_finalize):
Use GpmControl to notify GpmProfile when a suspend is about to happen,
or when it has just happened so the data valid flag is made false.
This stops oddball readings where the percentage time is used when
the machine is a sleep state. The profile code now waits for the start
of the next complete percentage point. Fixes #453853
Many thanks to Christopher Halse Rogers for the excellent debugging.

* test/Makefile.am:
Now we are using gpm-control in gpm-profile, pull it in as a dep.

svn path=/trunk/; revision=2321