Apply a patch from David Zeuthen: OK, so it turns out we're getting a
2007-10-05  Richard Hughes  <richard@hughsie.com>

* src/gpm-manager.c: (gpm_manager_action_suspend),
(gpm_manager_action_hibernate), (idle_do_sleep):
Apply a patch from David Zeuthen:
OK, so it turns out we're getting a D-Bus timeout. Actually g-p-m got
codepaths to handle this but the call site passes NULL for the
GError so we never get to examine this and properly handle timeouts.
This patch takes care of that. I've tested and it works. Fixes rh:312761


svn path=/trunk/; revision=2523