Fix laptops not hibernating on low battery
Do not recompute the percentage based on the charge when the percentage
reported by hal is less or equal than 1%. It seems hal doesn't computer
the percentage the same was as gnome-power-manager, and it can happen
that when we get the final event from hal, gnome-power-manager still
thinks it's too early for the critical action. However, since this was
the last event from hal, the critical action will never be triggered.

Fix bug #559578.