Revert "Add a bodge to reset the idletime XAlarm when we are no longer inhibited."
This reverts commit b5cf34d4b9861ab916a3ff74a2a08093ccb75b9f.

Peter has found and patched the bug in the X server that was causing the failure
of negative transition triggers in XSync.
For you and me, that means that XIDLETIME now works as expected, and gnome-session
and gnome-power-manager do the right thing. This means no more random blanking
when typing.

The important patch for the xserver is here:
http://cgit.freedesktop.org/xorg/xserver/patch/?id=db568f9eabf3450d8a023597ff007df355b13ea8