Avoid a race in detecting the idle reset alarm
When the idle timeout alarm fires, a reset alarm is created. If
the IDLETIME counter resets in between the idle timeout alarm being
detected and creating the reset alarm, then the reset alarm will
be missed and we will think that X is idle forever.

This patch creates a reset alarm when creating a EggIdletime
instance, but blocks emitting a "reset" signal from this instance
each time the reset alarm expires, unless the idle timeout alarm
has triggered since the last time the reset alarm expired. This
ensures that we always catch the reset alarm.