Fix a race condition where we can miss the idle reset alarm
When the idle timeout alarm fires, a reset alarm is created. If the IDLETIME counter resets in between the idle timeout alarm being detected and creating the reset alarm, then the reset alarm will be missed and we will think that X is idle forever.

This patch checks the current IDLETIME value from the server just after creating the reset alarm. If the current value is less than the reset threshold, then we assume that we missed the event and handle it accordingly.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=609720