Adjust dim timeout when necessary
On slow machines, or machines that have lots to load duing login,
the current idle time could be bigger than the requested timeout.
In this case the scheduled idle timeout will never fire, unless
some user activity (keyboard, mouse) resets the current idle time.
Instead of relying on user activity to correct this issue, we need
to adjust timeout, as related to current idle time, so the idle
timeout will fire as designed.

Signed-off-by: Richard Hughes <richard@hughsie.com>