If we loose the DBus name, exit
This prevents two instances of gnome-power-manager being run by gdm, which
itself causes an issue if the user closes the laptop lid, and both processes
tell upower to suspend.

One call to upower succeeds, and one fails, and the user is left with a
erroneous 'Your computer failed to suspend' message on resume.

In 3.2, gnome-power-manager is being moved to gnome-settings-daemon, and so the
problem goes away for good. Yay.