build: install binary into correct location
Before this commit build just expected that we're going
to install into ${PREFIX}/bin, which is wrong in some cases
(e.g. if you want to cross-compile and have to install shared
data into /usr/share and all arch specific stuff into /usr/${CTARGET})

By using get_option('bindir') we allow passing `--bindir` to meson
to customize it. If no `--bindir` param is passed to meson `bin`
is the default, so it doesn't change the install location if
nothing is passed to meson.